some clean up, need to see why the png doesnt' get served properly

+++ ray-tracer/src/main/scala/io/tuliplogic/raytracer/ops/programs/SimpleWorld.scala
@@ -4,27 +4,13 @@ import java.nio.file.{Path, Paths}
 
 import io.tuliplogic.raytracer.commons.errors.{ATError, RayTracerError}
 import io.tuliplogic.raytracer.geometry.affine.PointVec.{Pt, Vec}
-import io.tuliplogic.raytracer.geometry.affine.aTModule
 import io.tuliplogic.raytracer.geometry.affine.aTModule.ATModule
-import io.tuliplogic.raytracer.geometry.matrix.matrixModule
-import io.tuliplogic.raytracer.geometry.matrix.matrixModule.MatrixModule
 import io.tuliplogic.raytracer.ops.model.data.Scene.{PointLight, Shape, Sphere}
-import io.tuliplogic.raytracer.ops.model.data.rayModule.RayModule
-import io.tuliplogic.raytracer.ops.model.data.{Color, Material, World, rayModule}
-import io.tuliplogic.raytracer.ops.model.modules.cameraModule.CameraModule
-import io.tuliplogic.raytracer.ops.model.modules.lightReflectionModule.LightReflectionModule
-import io.tuliplogic.raytracer.ops.model.modules.phongReflectionModule.PhongReflectionModule
-import io.tuliplogic.raytracer.ops.model.modules.{cameraModule, lightDiffusionModule, lightReflectionModule, normalReflectModule, phongReflectionModule, rasteringModule, worldHitCompsModule, worldModule, worldReflectionModule, worldRefractionModule, worldTopologyModule}
+import io.tuliplogic.raytracer.ops.model.data.{Color, Material, World}
 import io.tuliplogic.raytracer.ops.model.modules.rasteringModule.RasteringModule
-import io.tuliplogic.raytracer.ops.model.modules.worldHitCompsModule.WorldHitCompsModule
-import io.tuliplogic.raytracer.ops.model.modules.worldModule.WorldModule
-import io.tuliplogic.raytracer.ops.model.modules.worldTopologyModule.WorldTopologyModule
 import io.tuliplogic.raytracer.ops.rendering.canvasSerializer
 import io.tuliplogic.raytracer.ops.rendering.canvasSerializer.CanvasSerializer
 import zio._
-import zio.blocking.Blocking
-import zio.clock.Clock
-import zio.scheduler.Scheduler
 
 object SimpleWorld extends App{
   val cameraFrom = Pt(2, 2, -18)
@@ -51,7 +37,7 @@ object SimpleWorld extends App{
   import layers._
 
   override def run(args: List[String]): ZIO[ZEnv, Nothing, Int] =
-    ZIO.traverse(-18 to -6)(z => program(Pt(2, 2, z), Paths.get(s"$canvasFile-$z.ppm"))
+    ZIO.traverse(-18 to -6)(z => program(Pt(2, 2, z.toDouble), Paths.get(s"$canvasFile-$z.ppm"))
       .provideLayer(cSerializerM ++ (atM >>> rasteringM) ++ atM)
     ).timed.foldM(err =>
     console.putStrLn(s"Execution failed with: $err").as(1),
